refactor: PostsManagerPage 구조 개선
