--- src/entities/post/lib/highlightText.ts
@@ -1,1 +1,0 @@
-// 텍스트 하이라이트 유틸
