+++ src/features/posts/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as PostsTable } from "./PostsTable"
