--- src/pages/PostsManagerPage/model/index.ts
@@ -1,1 +1,0 @@
-// 페이지 관련 Zustand 스토어
