--- src/pages/PostsManagerPage/ui/PostsFilters/index.tsx
@@ -1,1 +1,0 @@
-// 필터 및 정렬 UI
