--- src/pages/PostsManagerPage/ui/PostsPagination/index.tsx
@@ -1,1 +1,0 @@
-// 페이지네이션 UI
