--- src/pages/PostsManagerPage/ui/PostsSearchBar/index.tsx
@@ -1,1 +1,0 @@
-// 검색 관련 UI
