--- src/pages/PostsManagerPage/ui/PostsTable/index.tsx
@@ -1,1 +1,0 @@
-// 게시물 테이블 UI
